src/requirements.txt
click==8.1.3
cycler==0.11.0
dlib==19.23.1
fonttools==4.33.3
funcy==1.17
kiwisolver==1.4.2
matplotlib==3.5.2
nptyping==2.0.1
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
pycairo==1.21.0
PyGObject==3.42.1
pyparsing==3.0.8
python-dateutil==2.8.2
six==1.16.0
types-Pillow==9.0.14
